# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_satria_newproject_internal_services.CreateCommentRequest:
    properties:
      content:
        minLength: 1
        type: string
      parent_id:
        description: optional UUID string for nested replies
        type: string
    required:
    - content
    type: object
...
  github_com_satria_newproject_internal_services.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  github_com_satria_newproject_internal_services.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  github_com_satria_newproject_internal_services.RegisterRequest:
    properties:
      email:
        maxLength: 100
        type: string
      password:
        maxLength: 100
        minLength: 8
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  github_com_satria_newproject_internal_utils.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@socialforum.local
    name: API Support
  description: Backend API for social forum with threads, comments, likes, real-time
    chat, and image uploads.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Social Forum API
  version: 1.0.0
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate with email and password
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_satria_newproject_internal_services.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_satria_newproject_internal_utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_satria_newproject_internal_utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_satria_newproject_internal_utils.APIResponse'
      summary: Login user
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      description: Revoke all refresh tokens for current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_satria_newproject_internal_utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_satria_newproject_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - Auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Get new access token using refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_satria_newproject_internal_services.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_satria_newproject_internal_utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_satria_newproject_internal_utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_satria_newproject_internal_utils.APIResponse'
      summary: Refresh access token
      tags:
      - Auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Create account with username, email, and password
      parameters:
      - description: Registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_satria_newproject_internal_services.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_satria_newproject_internal_utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_satria_newproject_internal_utils.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_satria_newproject_internal_utils.APIResponse'
      summary: Register new user
      tags:
      - Auth
  /api/v1/comments/{id}:
    delete:
      consumes:
      - application/json
      description: Delete comment (owner only)
      parameters:
      - description: Comment ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_satria_newproject_internal_utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_satria_newproject_internal_utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_satria_newproject_internal_utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_satria_newproject_internal_utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_satria_newproject_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete comment
      tags:
      - Comments
...
  /api/v1/threads/{id}/chat:
    get:
      description: Upgrade to WebSocket for real-time chat in a thread
      parameters:
      - description: Thread ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: JWT token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_satria_newproject_internal_utils.APIResponse'
      summary: WebSocket chat
      tags:
      - Chat
  /api/v1/threads/{id}/comments:
    get:
      consumes:
      - application/json
      description: Get all comments for a thread (with nested replies)
      parameters:
      - description: Thread ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_satria_newproject_internal_utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_satria_newproject_internal_utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_satria_newproject_internal_utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_satria_newproject_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get comments
      tags:
      - Comments
    post:
      consumes:
      - application/json
      description: Create comment or nested reply on a thread
      parameters:
      - description: Thread ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Comment data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_satria_newproject_internal_services.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_satria_newproject_internal_utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_satria_newproject_internal_utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_satria_newproject_internal_utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_satria_newproject_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Create comment
      tags:
      - Comments
...
  /api/v1/threads/{id}/messages:
    get:
      consumes:
      - application/json
      description: Get message history for a thread
      parameters:
      - description: Thread ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - default: 50
        description: Number of messages
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_satria_newproject_internal_utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_satria_newproject_internal_utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_satria_newproject_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get chat messages
      tags:
      - Chat
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"